requirements.txt
# FLASK + PANEL (FULL)
Flask==3.0.3
Werkzeug==3.0.3
SQLAlchemy==1.4.54
python-dotenv==1.0.0
aioquic==1.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Cython==3.0.10
Flask==3.0.3
h11==0.16.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.5
mitmproxy==12.0.0
mitmproxy_rs==0.12.7
msgpack==1.0.8
passlib==1.7.4
protobuf==5.27.2
publicsuffix2==2.20191221
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
pydivert==2.1.0
pylsqpack==0.3.21
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.9.0
requests==2.31.0
discord.py==2.3.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
service-identity==24.1.0
setuptools==69.5.1
sortedcontainers==2.4.0
tornado==6.4.2
typing_extensions==4.14.0
urwid==2.6.16
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
zstandard==0.22.0
aiohttp==3.9.5